backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
python-multipart
langchain
langchain-google-genai
langchain-chroma
langchain-community
langchain-text-splitters